# packages/shared/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IAPVerifyRequest": {
        "properties": {
          "is_subscription": {
            "default": false,
            "title": "Is Subscription",
            "type": "boolean"
          },
          "platform": {
            "enum": [
              "apple",
              "google"
            ],
            "title": "Platform",
            "type": "string"
          },
          "product_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Product Id",
            "type": "string"
          },
          "purchase_token": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Token"
          },
          "receipt_data": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Data"
          },
          "transaction_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "product_id",
          "transaction_id"
        ],
        "title": "IAPVerifyRequest",
        "type": "object"
      },
      "IAPVerifyResponse": {
        "properties": {
          "credits_added": {
            "default": 0,
            "title": "Credits Added",
            "type": "integer"
          },
          "plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          },
          "verification_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Source"
          }
        },
        "required": [
          "status",
          "transaction_id",
          "product_id"
        ],
        "title": "IAPVerifyResponse",
        "type": "object"
      },
      "IAPWebhookRequest": {
        "properties": {
          "payload": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "status": {
            "enum": [
              "active",
              "cancelled",
              "expired",
              "refunded"
            ],
            "title": "Status",
            "type": "string"
          },
          "transaction_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "status"
        ],
        "title": "IAPWebhookRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "support@aistorybook.com",
      "name": "AI Story Book Team"
    },
    "description": "\n# AI Story Book API\n\nAI 기반 맞춤형 동화책 생성 API입니다.\n\n## 주요 기능\n\n* **책 생성**: 주제, 연령대, 스타일을 입력하면 AI가 동화책을 생성합니다\n* **캐릭터 관리**: 사진에서 캐릭터를 추출하거나 직접 생성할 수 있습니다\n* **시리즈 생성**: 같은 캐릭터로 연속된 이야기를 만들 수 있습니다\n* **PDF/오디오 내보내기**: 완성된 책을 PDF나 오디오로 내보낼 수 있습니다\n\n## 인증\n\n모든 API는 `X-User-Key` 헤더가 필요합니다.\n중복 요청 방지를 위해 `X-Idempotency-Key` 헤더 사용을 권장합니다.\n\n## Rate Limiting\n\n- 기본: 분당 10회 요청 제한\n- 초과 시 429 Too Many Requests 응답\n\n## 크레딧 시스템\n\n책 1권 생성에 크레딧 1개가 소모됩니다.\n    ",
    "license": {
      "name": "MIT"
    },
    "title": "AI Story Book API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/iap/verify": {
      "post": {
        "description": "결제 영수증 검증 및 보상 지급.",
        "operationId": "verify_iap_v1_iap_verify_post",
        "parameters": [
          {
            "description": "User identification key",
            "in": "header",
            "name": "x-user-key",
            "required": true,
            "schema": {
              "description": "User identification key",
              "title": "X-User-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IAPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IAPVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Iap",
        "tags": [
          "IAP"
        ]
      }
    },
    "/v1/iap/webhook/apple": {
      "post": {
        "operationId": "apple_webhook_v1_iap_webhook_apple_post",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Webhook-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IAPWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Webhook",
        "tags": [
          "IAP"
        ]
      }
    },
    "/v1/iap/webhook/google": {
      "post": {
        "operationId": "google_webhook_v1_iap_webhook_google_post",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Webhook-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IAPWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Webhook",
        "tags": [
          "IAP"
        ]
      }
    },
...
  },
...
}